Overhauled UI Setup
Previously, the UI was created as part of the HTML document, hard coded.
The script was given the ID's of the elements attached them to the
Google scripts and functions.

Now, the UI elements are created by the script. This is a full overhaul
of all of the UI, from the UI wrappers, to the UI elements inside those
wrappers.