Give addPolygon Function The Ability to Receive Custom Callback Function for Click Events
When the removal of the application specific logic that linked places to
polygons, I realized that there needed to be an easy way to add click
event listeners to polygons. This can be done manually in the
application script, but adding the ability to send the callback function
to the creation function reduces the required code in the application
script.