ShareLocation - Moved Google Setup into Main Public Function out of Separate Private Function
Previously, the logic for adding a share location button was split into
a public function, and a private function. The public function handled
creating the HTML elements, and the private function handled setting up
the Google logic.

However, the reason for a function is that it may be called multiple
times. The private function was only ever called directly by the public
function. It really had no need to exist. In fact it caused a lot of
duplicate logic These functions have been combined. This combining will
happen with the share location logic as well soon.