Changed The Way Custom Markers are Handled
The only reason to specify a custom marker as a global object is if it
should be the new default marker. That means that addMarker should use
the global unless overridden. It now does.

What this does is allow for a custom marker URL to be passed in at init,
and that becomes the default marker. When markers are added, they all
will use the default marker. However, if a markerURL is passed in as a
param directly to addMarker(), this can be overridden. Giving
markerURL:'' would cause that marker to be the normal Google marker.