Changed self.opts to Passed Parameter Instead of Global, and Renamed
The options passed to the function has been renamed to params, to match
convention throughout the rest of the script, and is no longer a
variable. This isolates the params in the init function, as they should
be. Global variables needed by the script are set up and validated in
init.