Improved Validation in Init
There are two required parameters for Mapstractor. mapWrapID is
required, and map is required. map is the object containing the options
for the Google map that is being created. mapWrapID is the ID of the
hardcoded HTML wrapper element. Additionally, it is required that there
is an actual hardcoded HTML element with an ID that matches the provided
mapWrapID.

All three of these conditions are now tested.