Added Ability to Clear All Infoboxes
This required adding a global array to track infoboxes. By doing this,
we can have multiple infoboxes on the map if desired.

Some interesting hackery was needed to allow the marker prototype
function to access the infobox list global. It required adding a link
from the main mapstractor object into the main gMap object, which the
marker has access to.