Replacing the mapWrapID API for mapWrapSelector
mapWrapID is limiting, because it requires a hard-coded ID on the page.
Sometimes users might be getting HTML from a CMS and not be in control
of IDs. Moving to a general selector gives much more flexibility.

For backwards-compatibility purposes, mapWrapID still works for now. It
won't in the future.